Relocate adventure, do some refactoring
