+++ pom.xml
@@ -56,12 +56,8 @@
                                     <shadedPattern>codes.wasabi.xclaim.paperlib</shadedPattern>
                                 </relocation>
                                 <relocation>
-                                    <pattern>net.kyori.adventure.platform</pattern>
-                                    <shadedPattern>codes.wasabi.xclaim.adventure.platform</shadedPattern>
-                                </relocation>
-                                <relocation>
-                                    <pattern>net.kyori.adventure.text.minimessage</pattern>
-                                    <shadedPattern>codes.wasabi.xclaim.adventure.minimessage</shadedPattern>
+                                    <pattern>net.kyori.adventure</pattern>
+                                    <shadedPattern>codes.wasabi.xclaim.adventure</shadedPattern>
                                 </relocation>
                             </relocations>
                         </configuration>
