+++ src/main/java/codes/wasabi/xclaim/platform/spigot_1_17/SpigotPlatformChatListener.java
@@ -1,4 +1,4 @@
-package codes.wasabi.xclaim.platform.spigot;
+package codes.wasabi.xclaim.platform.spigot_1_17;
 
 import codes.wasabi.xclaim.platform.PlatformChatListener;
 import org.bukkit.event.EventHandler;
